Switch SourceInfo from veneer to private macro interface
